Move the division tiers to src/divide.cpp
The Burnikel-Ziegler recursion (basecase, D_3n/2n, D_2n/n, the divappr
variant), the block drivers (exact, approximate, and quotient-only),
the exact Newton reciprocal, and the block Barrett driver are now
compiled once, taking only the type-erased scratch whose heap hooks
supply the internal products and FFT workspaces. divide_quotient's
padded numerator moves from an allocator-rebound vector into the limb
scratch, with burnikel_ziegler_plan_bits factored out so the new
divide_quotient_storage_size can size the padded plan without a
buffer.

The header keeps the constexpr schoolbook kernels, plans, storage
models, gates, dispatchers, and thin allocator-owning convenience
templates; those templates are constrained against deducing a scratch
type as the allocator (a derived scratch reference would otherwise
bind the template exactly and recurse) and forward through an explicit
base cast. Detail-level test callsites drop the allocator argument;
the public surfaces are unchanged.

Suites green on both configs (816/818), a 200k fuzz_div soak, and the
forced-32-bit standalone check (now linking src/divide.cpp et al).
